Move extractor in to skyit module
