+++ yt_dlp/extractor/_extractors.py
@@ -1882,6 +1882,8 @@ from .skyit import (
     SkyItVideoIE,
     SkyItVideoLiveIE,
     TV8ItIE,
+    TV8ItLiveIE,
+    TV8ItPlaylistIE,
 )
 from .skylinewebcams import SkylineWebcamsIE
 from .skynewsarabia import (
@@ -2192,7 +2194,6 @@ from .tv5unis import (
     TV5UnisIE,
     TV5UnisVideoIE,
 )
-from .tv8 import TV8PlaylistIE, TV8StreamingIE
 from .tv24ua import TV24UAVideoIE
 from .tva import TVAIE
 from .tvanouvelles import (
